+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- touchbundle package supports creation of metric bundles, which are logical groups of metrics
 
 ## [v0.1.0]
 - Updated go.uber.org/fx to 1.17.1